add evaluate_single; move wzksolver
