+++ wordle/core/__init__.py
@@ -1,6 +1,6 @@
 from .banker import WordleGame
 from .dictionary import WordleDictionary
 from .profiler import WordleProfiler
-from .solver import BaseWordleSolver, TheGreatWzkSolver
+from .solver import BaseWordleSolver
 
-__all__ = ["WordleGame", "WordleDictionary", "WordleProfiler", "BaseWordleSolver", "TheGreatWzkSolver"]
+__all__ = ["WordleGame", "WordleDictionary", "WordleProfiler", "BaseWordleSolver"]
